x
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@3682 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ kdc/connect.c
@@ -302,7 +302,8 @@ do_request(void *buf, size_t len,
 	krb5_data_free(&reply);
     }
     if(ret)
-	kdc_log(0, "Failed processing %d byte request from %s", len, addr);
+	kdc_log(0, "Failed processing %lu byte request from %s", 
+		(unsigned long)len, addr);
 }
 
 static void
